Refactor: Resolve logger circular dependency with DI
Merges `errorLogger.js` into a single `logger.js` and resolves the circular dependency with the configuration manager.

Previously, `logger.js` imported `configManager.js` to check the debug flag, while `configManager.js` imported `logger.js` to log errors, creating a circular dependency that was worked around with a separate `errorLogger.js` file.

This commit resolves the issue by using dependency injection:
- A `setDebug(boolean)` function is introduced in `logger.js`.
- `configManagerFactory.js` now imports and calls `setDebug` whenever the main config is loaded, updated, or reset.

This inverts the control flow, cleanly breaking the dependency loop and allowing for the consolidation of all logging functions into a single, more maintainable module. All import paths have been updated accordingly.
