+++ AddonExeBP/scripts/core/bountyManager.js
@@ -5,7 +5,7 @@
 
 import { world } from '@minecraft/server';
 import { debugLog } from './logger.js';
-import { errorLog } from './errorLogger.js';
+import { errorLog } from './logger.js';
 import * as playerDataManager from './playerDataManager.js';
 
 const bountyDataKey = 'exe:bountyData';
