+++ AddonExeBP/scripts/core/playerDataManager.js
@@ -25,7 +25,7 @@
 import { getConfig } from './configManager.js';
 import { world, system } from '@minecraft/server';
 import { debugLog } from './logger.js';
-import { errorLog } from './errorLogger.js';
+import { errorLog } from './logger.js';
 
 const playerPropertyPrefix = 'exe:player.';
 const playerNameIdMapKey = 'exe:playerNameIdMap';
