+++ AddonExeBP/scripts/core/punishmentManager.js
@@ -1,7 +1,7 @@
 import { world, system } from '@minecraft/server';
 import { getConfig } from './configManager.js';
 import { debugLog } from './logger.js';
-import { errorLog } from './errorLogger.js';
+import { errorLog } from './logger.js';
 
 const punishmentDbKey = 'exe:punishments';
