+++ AddonExeBP/scripts/core/shopManager.js
@@ -2,7 +2,7 @@ import { ItemStack, ItemTypes, EnchantmentTypes } from '@minecraft/server';
 import * as economyManager from './economyManager.js';
 import { getShopConfig } from './configurations.js';
 import { items as allItems } from './itemsConfig.js';
-import { errorLog } from './errorLogger.js';
+import { errorLog } from './logger.js';
 
 /**
  * Creates an ItemStack for a given item ID, handling enchantments.
